requirements.txt
asgiref==3.7.2
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.2
djangorestframework==3.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
stripe==5.4.0
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.3
